Divide ChunkManager into multiple classes
