+++ src/RealWorld/world/CMakeLists.txt
@@ -1,7 +1,10 @@
 ﻿add_subdirectory(shaders)
 
 target_sources(RealWorld PRIVATE 
+    ActivationManager.hpp       ActivationManager.cpp
+    ActiveChunksSB.hpp          
     Chunk.hpp                   Chunk.cpp
     ChunkManager.hpp            ChunkManager.cpp
+    VegManager.hpp              VegManager.cpp
     World.hpp                   World.cpp
 )
