+++ src/RealWorld/world/ChunkManager.cpp
@@ -1,20 +1,16 @@
 ﻿/*!
  *  @author    Dubsky Tomas
  */
-#include <algorithm>
-#include <array>
-#include <execution>
 #include <future>
 
 #include <RealEngine/graphics/synchronization/Fence.hpp>
 #include <RealEngine/utility/Math.hpp>
 
 #include <RealWorld/constants/vegetation.hpp>
-#include <RealWorld/save/ChunkLoader.hpp>
+#include <RealWorld/world/ActivationManager.hpp>
 #include <RealWorld/world/ChunkManager.hpp>
 
 using enum vk::BufferUsageFlagBits;
-using enum vk::MemoryPropertyFlagBits;
 using enum vk::ImageAspectFlagBits;
 using enum vk::CommandBufferUsageFlagBits;
 
@@ -23,101 +19,26 @@ using A = vk::AccessFlagBits2;
 
 namespace rw {
 
-vk::DeviceSize calcActiveChunksBufSize(glm::ivec2 worldTexSizeCh) {
-    glm::ivec2 maxContinuous = worldTexSizeCh - 1;
-    return sizeof(ChunkManager::ActiveChunksSB) +
-           sizeof(glm::ivec2) * (maxContinuous.x * maxContinuous.y +
-                                 worldTexSizeCh.x * worldTexSizeCh.y);
-}
-
-ChunkManager::ChunkManager(const re::PipelineLayout& pipelineLayout)
+ChunkManager::ChunkManager(
+    const re::PipelineLayout& pipelineLayout, ActivationManager& actManager
+)
     : m_tilesStageBuf(re::BufferCreateInfo{
           .allocFlags = vma::AllocationCreateFlagBits::eMapped |
                         vma::AllocationCreateFlagBits::eHostAccessRandom,
           .sizeInBytes = k_tileStageSize * k_chunkByteSize,
           .usage       = eTransferSrc | eTransferDst,
           .debugName   = "rw::ChunkManager::tilesStage"})
-    , m_analyzeContinuityPl(
-          {.pipelineLayout = *pipelineLayout,
-           .debugName      = "rw::ChunkManager::analyzeContinuity"},
-          {.comp = analyzeContinuity_comp}
-      )
-    , m_cullVegetationPl(
-          {.pipelineLayout = *pipelineLayout,
-           .debugName      = "rw::ChunkManager::cullVegetation"},
-          {.comp = saveVegetation_comp}
-      ) {
+    , m_actManager(actManager) {
 }
 
-const re::Buffer& ChunkManager::setTarget(const TargetInfo& targetInfo) {
-    m_folderPath = targetInfo.folderPath;
-    m_chunkGen.setTarget(ChunkGenerator::TargetInfo{
-        .seed           = targetInfo.seed,
-        .worldTex       = targetInfo.worldTex,
-        .worldTexSizeCh = targetInfo.worldTexSizeCh,
-        .bodiesBuf      = targetInfo.bodiesBuf,
-        .branchBuf      = targetInfo.branchBuf});
-    m_worldTex = &targetInfo.worldTex;
-
-    // Recalculate active chunks mask and analyzer dispatch size
-    m_worldTexSizeMask            = targetInfo.worldTexSizeCh - 1;
-    m_analyzeContinuityGroupCount = targetInfo.worldTexSizeCh / 8;
-
-    // Reset ActiveChunks storage buffer
-    vk::DeviceSize bufSize = calcActiveChunksBufSize(targetInfo.worldTexSizeCh);
-    m_activeChunksBuf      = re::Buffer{re::BufferCreateInfo{
-             .memoryUsage = vma::MemoryUsage::eAutoPreferDevice,
-             .sizeInBytes = bufSize,
-             .usage       = eStorageBuffer | eIndirectBuffer | eTransferDst,
-             .debugName   = "rw::ChunkManager::activeChunks"}};
-    m_activeChunksStageBuf = re::BufferMapped<ActiveChunksSB>{re::BufferCreateInfo{
-        .allocFlags = vma::AllocationCreateFlagBits::eMapped |
-                      vma::AllocationCreateFlagBits::eHostAccessRandom,
-        .sizeInBytes = bufSize,
-        .usage       = eTransferSrc,
-        .debugName   = "rw::ChunkManager::activeChunksStage"}};
-    m_activeChunksStageBuf->activeChunksMask  = m_worldTexSizeMask;
-    m_activeChunksStageBuf->worldTexSizeCh    = targetInfo.worldTexSizeCh;
-    m_activeChunksStageBuf->dynamicsGroupSize = glm::ivec4{0, 1, 1, 0};
-    int maxNumberOfUpdateChunks = m_worldTexSizeMask.x * m_worldTexSizeMask.y;
-    int lastChunkIndex          = maxNumberOfUpdateChunks +
-                         targetInfo.worldTexSizeCh.x * targetInfo.worldTexSizeCh.y;
-    for (int i = maxNumberOfUpdateChunks; i < lastChunkIndex; i++) {
-        m_activeChunksStageBuf->offsets[i] = k_chunkNotActive;
-    }
-    re::CommandBuffer::doOneTimeSubmit([&](const re::CommandBuffer& cmdBuf) {
-        // Copy whole buffer
-        vk::BufferCopy2 bufferCopy{0ull, 0ull, bufSize};
-        cmdBuf->copyBuffer2(vk::CopyBufferInfo2{
-            m_activeChunksStageBuf.buffer(), // Src buffer
-            m_activeChunksBuf.buffer(),      // Dst buffer
-            bufferCopy                       // Region
-        });
-    });
-    targetInfo.descriptorSet.write(
-        vk::DescriptorType::eStorageBuffer, 1u, 0u, m_activeChunksBuf, 0ull, bufSize
-    );
-
-    // Clear remnants of previous world
-    m_inactiveChunks.clear();
-    m_nextFreeTileStage = 0;
-
-    return m_activeChunksBuf;
+void ChunkManager::setTarget(glm::ivec2 worldTexSizeCh) {
+    m_worldTexSizeMask = worldTexSizeCh - 1;
+    resetTileStages();
 }
 
-bool ChunkManager::saveChunks() {
-    // Save all inactive chunks
-    std::for_each(
-        std::execution::par_unseq,
-        m_inactiveChunks.begin(),
-        m_inactiveChunks.end(),
-        [&](const auto& pair) {
-            saveChunk(pair.second.tiles().data(), pair.first);
-        }
-    );
-
+bool ChunkManager::saveChunks(const re::Texture& worldTex) {
     // Save all active chunks (they have to be downloaded)
-    assert(m_nextFreeTileStage == 0);
+    assert(hasFreeTransferSpace());
     re::CommandBuffer cmdBuf{{.debugName = "rw::ChunkManager::saveChunks"}};
     re::Fence         downloadFinishedFence{{}};
     cmdBuf->begin({eOneTimeSubmit});
@@ -129,29 +50,32 @@ bool ChunkManager::saveChunks() {
         A::eTransferRead,          // Dst access mask
         vk::ImageLayout::eGeneral, // Old image layout
         vk::ImageLayout::eGeneral, // New image layout
-        m_worldTex->image()
+        worldTex.image()
     );
     cmdBuf->pipelineBarrier2(vk::DependencyInfo{{}, {}, {}, imageBarrier});
 
     auto saveAllChunksInTileStage = [&]() {
         std::array<std::future<void>, k_tileStageSize> futures{};
-        for (size_t i = 0; i < m_nextFreeTileStage; ++i) {
+        for (size_t i = k_tileStageSize - 1; i > m_nextDownloadTileStage; --i) {
             futures[i] = std::async(std::launch::async, [this, i]() {
-                saveChunk(
-                    &m_tilesStageBuf[k_chunkByteSize * i], m_tileStageStates[i].posCh
+                m_actManager.saveChunk(
+                    &m_tilesStageBuf[k_chunkByteSize * i], m_tileStageTargetCh[i]
                 );
             });
         }
-        m_nextFreeTileStage = 0;
+        resetTileStages();
     };
 
     auto worldTexSize = m_worldTexSizeMask + 1;
     for (int y = 0; y < worldTexSize.y; ++y) {
         for (int x = 0; x < worldTexSize.x; ++x) {
-            auto posAc = glm::ivec2(x, y);
-            auto& activeChunk = activeChunkAtIndex(acToIndex(posAc, worldTexSize));
+            auto  posAc = glm::ivec2(x, y);
+            auto& activeChunk =
+                m_actManager.activeChunkAtIndex(acToIndex(posAc, worldTexSize));
             if (activeChunk != k_chunkNotActive) {
-                if (!planDownload(cmdBuf, activeChunk, chToTi(posAc))) { // If stage is full
+                if (hasFreeTransferSpace()) { // If there is space in the stage
+                    planDownload(cmdBuf, activeChunk, chToTi(posAc));
+                } else {
                     cmdBuf->end();
                     cmdBuf.submitToComputeQueue(*downloadFinishedFence);
                     downloadFinishedFence.wait();
@@ -170,7 +94,7 @@ bool ChunkManager::saveChunks() {
         {},                                // Dst access mask
         vk::ImageLayout::eGeneral,         // Old image layout
         vk::ImageLayout::eReadOnlyOptimal, // New image layout
-        m_worldTex->image()
+        worldTex.image()
     );
     cmdBuf->pipelineBarrier2(vk::DependencyInfo{{}, {}, {}, imageBarrier});
 
@@ -182,218 +106,88 @@ bool ChunkManager::saveChunks() {
     return true;
 }
 
-size_t ChunkManager::numberOfInactiveChunks() {
-    return m_inactiveChunks.size();
-}
-
-void ChunkManager::beginStep() {
-    // Check inactive chunks that have been inactive for too long
-    for (auto it = m_inactiveChunks.begin(); it != m_inactiveChunks.end();) {
-        // If the inactive chunk has not been used for a minute
-        if (it->second.step() >= k_physicsStepsPerSecond * 60) {
-            // Save the chunk to disk
-            saveChunk(it->second.tiles().data(), it->first);
-            // And remove it from the collection
-            it = m_inactiveChunks.erase(it);
-        } else {
-            it++;
-        }
-    }
-
-    // Finalize tile transfers from previous step
-    m_transparentChunkChanges = 0;
-    for (size_t i = 0; i < m_nextFreeTileStage; ++i) {
-        const auto& stageState = m_tileStageStates[i];
-        auto        posAc      = chToAc(stageState.posCh, m_worldTexSizeMask);
-        auto&       activeChunk =
-            activeChunkAtIndex(acToIndex(posAc, m_worldTexSizeMask + 1));
-        switch (stageState.transfer) {
-        case TileStageTransferState::Downloading:
-            // Copy the tiles aside from the stage
-            m_inactiveChunks.emplace(
-                stageState.posCh,
-                Chunk{stageState.posCh, &m_tilesStageBuf[k_chunkByteSize * i]}
-            );
-            // Signal that there is no active chunk at the spot
-            activeChunk = k_chunkNotActive;
-            break;
-        case TileStageTransferState::Uploading:
-            // Signal the new active chunk
-            activeChunk = stageState.posCh;
-            m_transparentChunkChanges++;
-            break;
-        }
-    }
-    m_nextFreeTileStage = 0;
-}
-
-void ChunkManager::planActivationOfChunks(
-    const re::CommandBuffer& cmdBuf,
-    glm::ivec2               botLeftTi,
-    glm::ivec2               topRightTi,
-    glm::uint                branchWriteBuf
-) {
-    glm::ivec2 botLeftCh  = tiToCh(botLeftTi);
-    glm::ivec2 topRightCh = tiToCh(topRightTi);
-
-    // Activate all chunks that at least partially overlap the area
-    for (int y = botLeftCh.y; y <= topRightCh.y; ++y) {
-        for (int x = botLeftCh.x; x <= topRightCh.x; ++x) {
-            planTransition(cmdBuf, glm::ivec2(x, y), branchWriteBuf);
-        }
-    }
-}
-
-int ChunkManager::endStep(const re::CommandBuffer& cmdBuf) {
-    // If there have been transparent changes
-    if (m_transparentChunkChanges > 0) {
-        // Reset the number of update chunks to zero
-        m_activeChunksStageBuf->dynamicsGroupSize.x = 0;
-
-        { // Wait for previous dispatch to finish
-            auto bufferBarrier = re::bufferMemoryBarrier(
-                S::eDrawIndirect | S::eComputeShader, // Src stage mask
-                A::eIndirectCommandRead | A::eShaderStorageRead, // Src access mask
-                S::eTransfer,      // Dst stage mask
-                A::eTransferWrite, // Dst access mask
-                m_activeChunksBuf.buffer()
-            );
-            cmdBuf->pipelineBarrier2({{}, {}, bufferBarrier, {}});
-        }
-
-        // Copy the update to active chunks buffer
-        auto texSizeCh   = m_worldTexSizeMask + 1;
-        auto copyRegions = std::to_array<vk::BufferCopy2>(
-            {vk::BufferCopy2{
-                 offsetof(ActiveChunksSB, dynamicsGroupSize),
-                 offsetof(ActiveChunksSB, dynamicsGroupSize),
-                 sizeof(m_activeChunksStageBuf->dynamicsGroupSize.x)},
-             vk::BufferCopy2{
-                 offsetof(ActiveChunksSB, offsets[0]) +
-                     sizeof(ActiveChunksSB::offsets[0]) * m_worldTexSizeMask.x *
-                         m_worldTexSizeMask.y,
-                 offsetof(ActiveChunksSB, offsets[0]) +
-                     sizeof(ActiveChunksSB::offsets[0]) * m_worldTexSizeMask.x *
-                         m_worldTexSizeMask.y,
-                 sizeof(glm::ivec2) * (texSizeCh.x * texSizeCh.y)}}
+int ChunkManager::beginStep() {
+    // Finalize uploads from previous step
+    for (size_t i = 0; i < m_nextUploadTileStage; ++i) {
+        glm::ivec2 posCh       = m_tileStageTargetCh[i];
+        auto       posAc       = chToAc(posCh, m_worldTexSizeMask);
+        auto&      activeChunk = m_actManager.activeChunkAtIndex(
+            acToIndex(posAc, m_worldTexSizeMask + 1)
         );
-        cmdBuf->copyBuffer2(vk::CopyBufferInfo2{
-            m_activeChunksStageBuf.buffer(), // Src buffer
-            m_activeChunksBuf.buffer(),      // Dst buffer
-            copyRegions                      // Regions
-        });
+        // Signal that the chunk is active
+        activeChunk = posCh;
+    }
 
-        { // Wait for the copy to finish
-            auto bufferBarrier = re::bufferMemoryBarrier(
-                S::eTransfer,      // Src stage mask
-                A::eTransferWrite, // Src access mask
-                S::eComputeShader, // Dst stage mask
-                A::eShaderStorageRead | A::eShaderStorageWrite, // Dst access mask
-                m_activeChunksBuf.buffer(),
-                offsetof(ActiveChunksSB, dynamicsGroupSize), // Offset
-                vk::WholeSize                                // Size
-            );
-            cmdBuf->pipelineBarrier2({{}, {}, bufferBarrier, {}});
-        }
-
-        // Analyze the world texture
-        cmdBuf->bindPipeline(vk::PipelineBindPoint::eCompute, *m_analyzeContinuityPl);
-        cmdBuf->dispatch(
-            m_analyzeContinuityGroupCount.x, m_analyzeContinuityGroupCount.y, 1
+    // Finalize downloads from previous step
+    for (size_t i = k_tileStageSize - 1; i > m_nextDownloadTileStage; --i) {
+        glm::ivec2 posCh       = m_tileStageTargetCh[i];
+        auto       posAc       = chToAc(posCh, m_worldTexSizeMask);
+        auto&      activeChunk = m_actManager.activeChunkAtIndex(
+            acToIndex(posAc, m_worldTexSizeMask + 1)
         );
 
-        // Save vegetation
-        // cmdBuf->bindPipeline(vk::PipelineBindPoint::eCompute, *m_cullVegetationPl);
-        // cmdBuf->dispatch(re::ceilDiv(k_maxVegCount, 256), 1, 1);
+        // Copy the tiles aside from the stage
+        m_actManager.addInactiveChunk(
+            posCh, Chunk{posCh, &m_tilesStageBuf[k_chunkByteSize * i]}
+        );
 
-        { // Barrier analysis output from tile transformations
-            auto bufferBarrier = re::bufferMemoryBarrier(
-                S::eComputeShader,       // Src stage mask
-                A::eShaderStorageWrite,  // Src access mask
-                S::eDrawIndirect,        // Dst stage mask
-                A::eIndirectCommandRead, // Dst access mask
-                m_activeChunksBuf.buffer()
-            );
-            cmdBuf->pipelineBarrier2({{}, {}, bufferBarrier, {}});
-        }
+        // Signal that there is no active chunk at the spot
+        activeChunk = k_chunkNotActive;
     }
-    return m_transparentChunkChanges;
+
+    int nTransparentChanges = m_nextUploadTileStage;
+    resetTileStages();
+    return nTransparentChanges;
 }
 
-void ChunkManager::planTransition(
-    const re::CommandBuffer& cmdBuf, glm::ivec2 posCh, glm::uint branchWriteBuf
-) {
-    auto posAc = chToAc(posCh, m_worldTexSizeMask);
-    auto& activeChunk = activeChunkAtIndex(acToIndex(posAc, m_worldTexSizeMask + 1));
-    if (activeChunk == posCh) {
-        // Chunk has already been active
-        return; // No transition is needed
-    } else if (activeChunk == k_chunkNotActive) {
-        // No chunk is active at the spot
-        planActivation(cmdBuf, activeChunk, posCh, chToTi(posAc), branchWriteBuf);
-    } else if (activeChunk != k_chunkBeingDownloaded && activeChunk != k_chunkBeingUploaded) {
-        // A different chunk is active at the spot
-        planDeactivation(cmdBuf, activeChunk, chToTi(posAc));
-    }
+bool ChunkManager::hasFreeTransferSpace() const {
+    return m_nextUploadTileStage < m_nextDownloadTileStage;
 }
 
-void ChunkManager::planActivation(
-    const re::CommandBuffer& cmdBuf,
-    glm::ivec2&              activeChunk,
-    glm::ivec2               posCh,
-    glm::ivec2               posAt,
-    glm::uint                branchWriteBuf
+void ChunkManager::planUpload(
+    const std::vector<unsigned char>& tiles, glm::ivec2 posCh, glm::ivec2 posAt
 ) {
-    // Try to find the chunk among inactive chunks
-    auto it = m_inactiveChunks.find(posCh);
-    if (it != m_inactiveChunks.end()) {
-        // Query upload of the chunk
-        if (planUpload(cmdBuf, it->second.tiles(), posCh, posAt)) {
-            // Remove the chunk from inactive chunks
-            m_inactiveChunks.erase(it);
-            // And signal that it is being uploaded
-            activeChunk = k_chunkBeingUploaded;
-        }
-    } else {
-        auto tiles = ChunkLoader::loadChunk(m_folderPath, posCh, iChunkTi);
-        if (tiles.size() > 0) { // If chunk has been loaded
-            if (planUpload(cmdBuf, tiles, posCh, posAt)) {
-                // Signal that it is being uploaded
-                activeChunk = k_chunkBeingUploaded;
-            } else {
-                // Could not upload the chunk
-                // At least store it as an inactive chunk
-                m_inactiveChunks.emplace(posCh, Chunk{posCh, std::move(tiles)});
-            }
-        } else {
-            // Chunk is not on the disk, it has to be generated
-            m_chunkGen.generateChunk(
-                cmdBuf,
-                ChunkGenerator::OutputInfo{.posCh = posCh, .branchWriteBuf = branchWriteBuf}
-            );
-            activeChunk = posCh;
-            m_transparentChunkChanges++;
-        }
-    }
+    assert(hasFreeTransferSpace());
+
+    m_tileStageTargetCh[m_nextUploadTileStage] = posCh;
+    auto bufOffset = static_cast<vk::DeviceSize>(m_nextUploadTileStage) *
+                     k_chunkByteSize;
+    std::memcpy(&m_tilesStageBuf[bufOffset], tiles.data(), k_chunkByteSize);
+
+    m_copyRegions[m_nextUploadTileStage] = vk::BufferImageCopy2{
+        bufOffset, // Buffer offset
+        0u,
+        0u,                          // Tightly packed
+        {eColor, 0u, 0u, 1u},        // Subresource
+        {posAt.x, posAt.y, 0u},      // Offset
+        {iChunkTi.x, iChunkTi.y, 1u} // Extent
+    };
+    m_nextUploadTileStage++;
 }
 
-void ChunkManager::planDeactivation(
-    const re::CommandBuffer& cmdBuf, glm::ivec2& activeChunk, glm::ivec2 posAt
+void ChunkManager::planDownload(
+    const re::CommandBuffer& cmdBuf, glm::ivec2 posCh, glm::ivec2 posAt
 ) {
-    // Query download of the chunk
-    if (planDownload(cmdBuf, activeChunk, posAt)) {
-        activeChunk = k_chunkBeingDownloaded; // Deactivate the spot
-        m_transparentChunkChanges++;
-    }
+    assert(hasFreeTransferSpace());
+
+    m_tileStageTargetCh[m_nextDownloadTileStage] = posCh;
+    auto bufOffset = static_cast<vk::DeviceSize>(m_nextDownloadTileStage) *
+                     k_chunkByteSize;
+
+    m_copyRegions[m_nextDownloadTileStage] = vk::BufferImageCopy2{
+        bufOffset, // Buffer offset
+        0u,
+        0u,                          // Tightly packed
+        {eColor, 0u, 0u, 1u},        // Subresource
+        {posAt.x, posAt.y, 0u},      // Offset
+        {iChunkTi.x, iChunkTi.y, 1u} // Extent
+    };
+    m_nextDownloadTileStage--;
 }
 
-bool ChunkManager::planUpload(
-    const re::CommandBuffer&          cmdBuf,
-    const std::vector<unsigned char>& tiles,
-    glm::ivec2                        posCh,
-    glm::ivec2                        posAt
-) {
-    if (m_nextFreeTileStage < k_tileStageSize) { // If there is a free stage
+void ChunkManager::endStep(const re::CommandBuffer& cmdBuf, const re::Texture& worldTex) {
+    // If there are uploads scheduled
+    if (m_nextUploadTileStage > 0) {
         // Wait for unrasterization to finish
         auto imageBarrier = re::imageMemoryBarrier(
             S::eColorAttachmentOutput | S::eTransfer, // Src stage mask
@@ -403,38 +197,21 @@ bool ChunkManager::planUpload(
             A::eTransferWrite,                        // Dst access mask
             vk::ImageLayout::eGeneral,                // Old image layout
             vk::ImageLayout::eGeneral,                // New image layout
-            m_worldTex->image()
+            worldTex.image()
         );
         cmdBuf->pipelineBarrier2(vk::DependencyInfo{{}, {}, {}, imageBarrier});
 
-        m_tileStageStates[m_nextFreeTileStage] = {
-            .transfer = TileStageTransferState::Uploading, .posCh = posCh};
-        auto bufOffset = static_cast<vk::DeviceSize>(m_nextFreeTileStage) *
-                         k_chunkByteSize;
-        std::memcpy(&m_tilesStageBuf[bufOffset], tiles.data(), k_chunkByteSize);
-        auto copy = vk::BufferImageCopy2{
-            bufOffset, // Buffer offset
-            0u,
-            0u,                          // Tightly packed
-            {eColor, 0u, 0u, 1u},        // Subresource
-            {posAt.x, posAt.y, 0u},      // Offset
-            {iChunkTi.x, iChunkTi.y, 1u} // Extent
-        };
+        std::span spanOfRegions{
+            m_copyRegions.begin(), m_copyRegions.begin() + m_nextUploadTileStage};
         cmdBuf->copyBufferToImage2(vk::CopyBufferToImageInfo2{
             m_tilesStageBuf.buffer(),
-            m_worldTex->image(),
+            worldTex.image(),
             vk::ImageLayout::eGeneral,
-            copy});
-        m_nextFreeTileStage++;
-        return true; // Upload querried
+            spanOfRegions});
     }
-    return false; // No free stage to do the transfer
-}
 
-bool ChunkManager::planDownload(
-    const re::CommandBuffer& cmdBuf, glm::ivec2 posCh, glm::ivec2 posAt
-) {
-    if (m_nextFreeTileStage < k_tileStageSize) { // If there is a free stage
+    // If there are downloads scheduled
+    if (m_nextDownloadTileStage < k_tileStageSize - 1) {
         // Wait for unrasterization to finish
         auto imageBarrier = re::imageMemoryBarrier(
             S::eColorAttachmentOutput | S::eTransfer, // Src stage mask
@@ -444,39 +221,24 @@ bool ChunkManager::planDownload(
             A::eTransferRead,                         // Dst access mask
             vk::ImageLayout::eGeneral,                // Old image layout
             vk::ImageLayout::eGeneral,                // New image layout
-            m_worldTex->image()
+            worldTex.image()
         );
         cmdBuf->pipelineBarrier2(vk::DependencyInfo{{}, {}, {}, imageBarrier});
 
-        m_tileStageStates[m_nextFreeTileStage] = {
-            .transfer = TileStageTransferState::Downloading, .posCh = posCh};
-        auto copy = vk::BufferImageCopy2{
-            static_cast<vk::DeviceSize>(m_nextFreeTileStage) *
-                k_chunkByteSize, // Buffer offset
-            0u,
-            0u,                          // Tightly packed
-            {eColor, 0u, 0u, 1u},        // Subresource
-            {posAt.x, posAt.y, 0u},      // Offset
-            {iChunkTi.x, iChunkTi.y, 1u} // Extent
-        };
+        std::span spanOfRegions{
+            m_copyRegions.begin() + m_nextDownloadTileStage + 1,
+            m_copyRegions.end()};
         cmdBuf->copyImageToBuffer2(vk::CopyImageToBufferInfo2{
-            m_worldTex->image(),
+            worldTex.image(),
             vk::ImageLayout::eGeneral,
             m_tilesStageBuf.buffer(),
-            copy});
-        m_nextFreeTileStage++;
-        return true; // Download querried
+            spanOfRegions});
     }
-    return false; // No free stage to do the transfer
 }
 
-void ChunkManager::saveChunk(const uint8_t* tiles, glm::ivec2 posCh) const {
-    ChunkLoader::saveChunk(m_folderPath, posCh, iChunkTi, tiles);
-}
-
-glm::ivec2& ChunkManager::activeChunkAtIndex(int acIndex) {
-    return m_activeChunksStageBuf
-        ->offsets[m_worldTexSizeMask.x * m_worldTexSizeMask.y + acIndex];
+void ChunkManager::resetTileStages() {
+    m_nextUploadTileStage   = 0;
+    m_nextDownloadTileStage = k_tileStageSize - 1;
 }
 
 } // namespace rw
